requirements.txt
anyio==4.12.0
certifi==2025.11.12
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pymongo==4.15.5
python-dotenv==1.2.1
python-telegram-bot==22.5
